fix: import contract explicitly, which will cause conflict with InscriptionFactoryUpgrade.t.sol
import { InscriptionFactoryV1 } from '../src/Upgrade/InscriptionFactoryV1.sol'; is an explicit import that only brings in the InscriptionFactoryV1 contract or library, avoiding unnecessary imports and potential naming conflicts.
On the other hand, import '../src/upgrade/InscriptionFactoryV1.sol'; is an implicit import that includes all definitions from the file, which might lead to naming conflicts if multiple contracts or libraries are defined in the file.
Explicit imports are recommended for clarity and safety.